Move common Web integration tests features to web library file
